AR::Base.table_exists? returns false if called while no tenant is set.
This prevents errors during boot-time eager loading when trying to
load the schema (for the schema cache).

It feels really ugly to do this, and I'm sure there are more methods
which _should_ be stubbed like this to handle the no-tenant case, or a
simpler fix that addresses the general problem, but this is the only
method in my way right now and it moves me forward.