config.active_record_tenanted.connection_class
is a config option that takes a class name (default
"ApplicationRecord") that this gem will integrate into Rails test
setup and Record classes (to share a connection pool with that
connection class and create/use tables in that class's database). Or
set to `nil` to turn off integrations.

Also introduce config.active_record_tenanted.tenant_rails_records
which defaults to `true` but can be set to `false` to leave Record
classes in an untenanted primary database.

Also added an `after_initialize` block into the railtie to apply the
connection class configs.