ActiveRecord::Tenanted.connection_class
is an attribute that defaults to "ApplicationRecord" but can be set to
another class name to make Rails integrate testing and AR records with
that connection class. Or set to `nil` to turn off integrations.

Also introduce Tenanted.tenant_rails_records which defaults to `true`
but can be set to `false` to leave AR records in an untenanted primary
database.

Also added an `after_intialize` block into the railtie to apply the
connection class configs.